move term agg tests back into single file and added TODO for better TermsIncludeExluce type in 6.0
