--- src/Tests/Aggregations/Bucket/Terms/NumericTermsAggregationUsageTests.cs
@@ -1,63 +1,0 @@
-using System;
-using FluentAssertions;
-using Nest;
-using Tests.Framework.Integration;
-using Tests.Framework.MockData;
-
-namespace Tests.Aggregations.Bucket.Terms
-{
-	public class NumericTermsAggregationUsageTests : AggregationUsageTestBase
-	{
-		public NumericTermsAggregationUsageTests(ReadOnlyCluster i, EndpointUsage usage) : base(i, usage) { }
-
-		protected override object ExpectJson => new
-		{
-			size = 0,
-			aggs = new
-			{
-				commits = new
-				{
-
-					terms = new
-					{
-						field = "numberOfCommits"
-					}
-				}
-			}
-		};
-
-		protected override Func<SearchDescriptor<Project>, ISearchRequest> Fluent => s => s
-			.Size(0)
-			.Aggregations(a => a
-				.Terms("commits", st => st
-					.Field(p => p.NumberOfCommits)
-				)
-			);
-
-		protected override SearchRequest<Project> Initializer =>
-			new SearchRequest<Project>
-			{
-				Size = 0,
-				Aggregations = new TermsAggregation("commits")
-				{
-					Field = Infer.Field<Project>(p => p.NumberOfCommits),
-				}
-			};
-
-		protected override void ExpectResponse(ISearchResponse<Project> response)
-		{
-			response.IsValid.Should().BeTrue();
-			var commits = response.Aggs.Terms<int>("commits");
-			commits.Should().NotBeNull();
-			commits.DocCountErrorUpperBound.Should().HaveValue();
-			commits.SumOtherDocCount.Should().HaveValue();
-			commits.Buckets.Should().NotBeNull();
-			commits.Buckets.Count.Should().BeGreaterThan(0);
-			foreach (var item in commits.Buckets)
-			{
-				item.Key.Should().BeGreaterThan(0);
-				item.DocCount.Should().BeGreaterOrEqualTo(1);
-			}
-		}
-	}
-}
