+++ src/Tests/Tests.csproj
@@ -205,8 +205,6 @@
     <Compile Include=".\Document\Single\Update\UpdateApiTests.cs" />
     <Compile Include=".\Document\Single\Update\UpdateUrlTests.cs" />
     <Compile Include=".\Document\Single\Update\UpdateWithSourceApiTests.cs" />
-    <Compile Include="Aggregations\Bucket\Terms\NumericTermsAggregationUsageTests.cs" />
-    <Compile Include="Aggregations\Bucket\Terms\PartitionTermsAggregationUsageTests.cs" />
     <Compile Include="Aggregations\Matrix\MatrixStats\MatrixStatsAggregationUsageTests.cs" />
     <Compile Include="Aggregations\Metric\GeoCentroid\GeoCentroidAggregationUsageTests.cs" />
     <Compile Include="Aggregations\ReservedAggNames.doc.cs" />
